#1559 remove travis, move testing.md, shift constants
