--- .travis.yml
@@ -1,22 +1,0 @@
-language: node_js
-node_js:
-  - "10"
-cache:
-  yarn: true
-  directories:
-    - node_modules
-before_install:
-  - curl -o- -L https://yarnpkg.com/install.sh | bash
-  - export PATH=$HOME/.yarn/bin:$PATH
-script:
-  - yarn;yarn run build
-deploy:
-  provider: pages
-  keep-history: true
-  skip_cleanup: true
-  local_dir: dist/
-  github-token: $GITHUB_TOKEN
-  repo: dysonring/vue-apps-polkadot
-  target-branch: gh-pages
-  on:
-    branch: master
