+++ components/series/SeriesTable.vue
@@ -296,7 +296,7 @@ import { columns, nftFn } from './utils'
 import collectionSeriesList from '@/queries/collectionSeriesList.graphql'
 import { seriesAggQuery } from '../rmrk/Gallery/Search/query'
 import { NFTMetadata, Collection } from '../rmrk/service/scheme'
-import { denyList } from '@/constants'
+import { denyList } from '@/utils/constants'
 import { sanitizeIpfsUrl, fetchCollectionMetadata} from '@/components/rmrk/utils'
 import { exist } from '@/components/rmrk/Gallery/Search/exist'
 import { emptyObject } from '@/utils/empty'
