+++ components/spotlight/SpotlightTable.vue
@@ -132,7 +132,7 @@ import { columns, nftFn } from './utils'
 import collectionIssuerList from '@/queries/collectionIssuerList.graphql'
 import { spotlightAggQuery } from '../rmrk/Gallery/Search/query'
 import TransactionMixin from '@/utils/mixins/txMixin'
-import { denyList } from '@/constants'
+import { denyList } from '@/utils/constants'
 import { GenericAccountId } from '@polkadot/types/generic/AccountId'
 import { get } from 'idb-keyval'
 import { identityStore } from '@/utils/idbStore'
